Add files via upload
I have created a simple web based (local host) called "Brain Game" as it tests our memory and remembering power. we have to select the colour that it flashes and continues as a series. Once the user gets it wrong it will flash a red colour and hence the game will restart all over again. The user will be scored on the basis of number of levels user passed.

I used HTML, CSS, JavaScript to create this mini project